nvme: move resv commands into a plugin
There are too many top-level commands cluttering the space. Move the
reservation commands into a plugin, e.g.,

  nvme resv-acquire -> nvme resv acquire
